refactor: collapse src/components/home/ into home-page/
src/components/home/ and src/components/home-page/ both existed.
home-page/ is the canonical set imported by src/app/home-page/index.tsx.
Auditing home/ found:

- Testimonials was the only used component (imported once)
- Contactus, HeroSection, MissionSection, Ourblogs, SupportFreeforCharity
  had no inbound imports and were orphaned legacy code

Migrates Testimonials to src/components/home-page/Testimonials/
(updates the single import site in src/app/home-page/index.tsx)
and deletes the five orphan directories. The home/ directory now no
longer exists.

Refs #288
